♻️ refactor: 쿼리 생성 함수의 파일명 변경
