+++ src/domains/randomDefense/randomDefenseQueryGenerator.ts
